requirements.txt
asgiref==3.7.2
boto3==1.34.4
botocore==1.34.4
Django==4.2
django-environ==0.11.2
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
gunicorn==21.2.0
jmespath==1.0.1
Markdown==3.5.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.9.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
